+++ src/backend/uploader/script/ScriptUploader.h
@@ -28,19 +28,20 @@
 #include "src/backend/uploader/script/IScriptUploader.h"
 #include "src/backend/config/IConfig.h"
 #include "src/common/enum/UploadStatus.h"
-#include "src/common/provider/TempFileProvider.h"
+#include "src/common/provider/ITempFileProvider.h"
 
 class ScriptUploader : public IScriptUploader
 {
 	Q_OBJECT
 public:
-	explicit ScriptUploader(const QSharedPointer<IConfig> &config);
+	explicit ScriptUploader(const QSharedPointer<IConfig> &config, const QSharedPointer<ITempFileProvider> &tempFileProvider);
 	~ScriptUploader() override = default;
 	void upload(const QImage &image) override;
 	UploaderType type() const override;
 
 private:
 	QSharedPointer<IConfig> mConfig;
+    QSharedPointer<ITempFileProvider> mTempFileProvider;
 	QProcess mProcessHandler;
 	QString mPathToTmpImage;
 
@@ -48,7 +49,6 @@ private slots:
 	void scriptFinished(int exitCode, QProcess::ExitStatus exitStatus);
 	void errorOccurred(QProcess::ProcessError error);
 	bool saveImageLocally(const QImage &image);
-	void cleanup();
 	QString parseOutput(const QString &output) const;
 	void writeToConsole(const QString &output) const;
 	UploadStatus mapErrorTypeToStatus(QProcess::ProcessError errorType) const;